+++ benches/trace.rs
@@ -64,13 +64,13 @@ fn criterion_benchmark(c: &mut Criterion) {
     });
 }
 
-fn trace_benchmark_group<F: Fn(&sdk::Tracer) -> ()>(c: &mut Criterion, name: &str, f: F) {
+fn trace_benchmark_group<F: Fn(&sdk::Tracer)>(c: &mut Criterion, name: &str, f: F) {
     let mut group = c.benchmark_group(name);
 
     group.bench_function("always-sample", |b| {
         let always_sample = sdk::Provider::builder()
             .with_config(sdk::Config {
-                default_sampler: Box::new(sdk::Sampler::Always),
+                default_sampler: Box::new(sdk::Sampler::AlwaysOn),
                 ..Default::default()
             })
             .build()
@@ -82,7 +82,7 @@ fn trace_benchmark_group<F: Fn(&sdk::Tracer) -> ()>(c: &mut Criterion, name: &st
     group.bench_function("never-sample", |b| {
         let never_sample = sdk::Provider::builder()
             .with_config(sdk::Config {
-                default_sampler: Box::new(sdk::Sampler::Never),
+                default_sampler: Box::new(sdk::Sampler::AlwaysOff),
                 ..Default::default()
             })
             .build()