+++ examples/actix-udp/src/main.rs
@@ -1,0 +1,61 @@
+use actix_service::Service;
+use actix_web::middleware::Logger;
+use actix_web::{web, App, HttpServer};
+use opentelemetry::api::trace::futures::FutureExt;
+use opentelemetry::api::{Key, TraceContextExt, Tracer};
+use opentelemetry::{global, sdk};
+
+fn init_tracer() -> thrift::Result<()> {
+    let exporter = opentelemetry_jaeger::Exporter::builder()
+        .with_agent_endpoint("127.0.0.1:6831".parse().unwrap())
+        .with_process(opentelemetry_jaeger::Process {
+            service_name: "trace-demo".to_string(),
+            tags: vec![
+                Key::new("exporter").string("jaeger"),
+                Key::new("float").f64(312.23),
+            ],
+        })
+        .init()?;
+    let provider = sdk::Provider::builder()
+        .with_simple_exporter(exporter)
+        .with_config(sdk::Config {
+            default_sampler: Box::new(sdk::Sampler::AlwaysOn),
+            ..Default::default()
+        })
+        .build();
+    global::set_provider(provider);
+
+    Ok(())
+}
+
+async fn index() -> &'static str {
+    let tracer = global::tracer("request");
+    tracer.in_span("index", |ctx| {
+        ctx.span().set_attribute(Key::new("parameter").i64(10));
+        "Index"
+    })
+}
+
+#[actix_rt::main]
+async fn main() -> std::io::Result<()> {
+    std::env::set_var("RUST_LOG", "debug");
+    env_logger::init();
+    init_tracer().expect("Failed to initialise tracer.");
+
+    HttpServer::new(|| {
+        App::new()
+            .wrap(Logger::default())
+            .wrap_fn(|req, srv| {
+                let tracer = global::tracer("request");
+                tracer.in_span("middleware", move |cx| {
+                    cx.span().set_attribute(Key::new("path").string(req.path()));
+                    srv.call(req).with_context(cx)
+                })
+            })
+            .route("/", web::get().to(index))
+    })
+    .bind("127.0.0.1:8088")
+    .unwrap()
+    .run()
+    .await
+}