+++ examples/grpc/src/client.rs
@@ -4,7 +4,7 @@ use opentelemetry::api::{
     Context, HttpTextFormat, KeyValue, TraceContextExt, TraceContextPropagator, Tracer,
 };
 use opentelemetry::sdk::Sampler;
-use opentelemetry::{api, global, sdk};
+use opentelemetry::{global, sdk};
 
 pub mod hello_world {
     tonic::include_proto!("helloworld");
@@ -21,12 +21,12 @@ fn tracing_init() -> Result<(), Box<dyn std::error::Error>> {
         })
         .init()?;
 
-    // For the demonstration, use `Sampler::Always` sampler to sample all traces. In a production
-    // application, use `Sampler::Parent` or `Sampler::Probability` with a desired probability.
+    // For the demonstration, use `Sampler::AlwaysOn` sampler to sample all traces. In a production
+    // application, use `Sampler::ParentOrElse` or `Sampler::Probability` with a desired probability.
     let provider = sdk::Provider::builder()
         .with_simple_exporter(exporter)
         .with_config(sdk::Config {
-            default_sampler: Box::new(Sampler::Always),
+            default_sampler: Box::new(Sampler::AlwaysOn),
             ..Default::default()
         })
         .build();
@@ -34,22 +34,6 @@ fn tracing_init() -> Result<(), Box<dyn std::error::Error>> {
     global::set_provider(provider);
 
     Ok(())
-}
-
-struct TonicMetadataMapCarrier<'a>(&'a mut tonic::metadata::MetadataMap);
-impl<'a> api::Carrier for TonicMetadataMapCarrier<'a> {
-    fn get(&self, key: &'static str) -> Option<&str> {
-        self.0.get(key).and_then(|metadata| metadata.to_str().ok())
-    }
-
-    fn set(&mut self, key: &'static str, value: String) {
-        if let Ok(key) = tonic::metadata::MetadataKey::from_bytes(key.to_lowercase().as_bytes()) {
-            self.0.insert(
-                key,
-                tonic::metadata::MetadataValue::from_str(&value).unwrap(),
-            );
-        }
-    }
 }
 
 #[tokio::main]
@@ -63,7 +47,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut request = tonic::Request::new(HelloRequest {
         name: "Tonic".into(),
     });
-    propagator.inject_context(&cx, &mut TonicMetadataMapCarrier(request.metadata_mut()));
+    propagator.inject_context(&cx, request.metadata_mut());
 
     let response = client.say_hello(request).await?;