+++ examples/tracing-grpc/src/server.rs
@@ -1,0 +1,103 @@
+use hello_world::greeter_server::{Greeter, GreeterServer};
+use hello_world::{HelloReply, HelloRequest};
+use opentelemetry::api::{self, HttpTextFormat, KeyValue, Provider};
+use opentelemetry::sdk::{self, Sampler};
+use tonic::{transport::Server, Request, Response, Status};
+use tracing::*;
+use tracing_opentelemetry::OpenTelemetrySpanExt;
+use tracing_subscriber::prelude::*;
+
+pub mod hello_world {
+    tonic::include_proto!("helloworld"); // The string specified here must match the proto package name
+}
+
+#[instrument]
+fn expensive_fn(to_print: String) {
+    for _ in 0..5 {
+        std::thread::sleep(std::time::Duration::from_secs(1));
+        info!("{}", to_print);
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct MyGreeter {}
+
+#[tonic::async_trait]
+impl Greeter for MyGreeter {
+    #[instrument]
+    async fn say_hello(
+        &self,
+        request: Request<HelloRequest>, // Accept request of type HelloRequest
+    ) -> Result<Response<HelloReply>, Status> {
+        let propagator = api::TraceContextPropagator::new();
+        let parent_cx = propagator.extract(&HttpHeaderMapCarrier(request.metadata()));
+        let span = tracing::Span::current();
+        span.set_parent(&parent_cx);
+        let name = request.into_inner().name;
+        expensive_fn(format!("Got name: {:?}", name));
+
+        // Return an instance of type HelloReply
+        let reply = hello_world::HelloReply {
+            message: format!("Hello {}!", name), // We must use .into_inner() as the fields of gRPC requests and responses are private
+        };
+
+        Ok(Response::new(reply)) // Send back our formatted greeting
+    }
+}
+
+fn tracing_init() -> Result<(), Box<dyn std::error::Error>> {
+    let builder = opentelemetry_jaeger::Exporter::builder()
+        .with_agent_endpoint("127.0.0.1:6831".parse().unwrap());
+
+    let exporter = builder
+        .with_process(opentelemetry_jaeger::Process {
+            service_name: "grpc-server".to_string(),
+            tags: vec![KeyValue::new("version", "0.1.0")],
+        })
+        .init()?;
+
+    // For the demonstration, use `Sampler::Always` sampler to sample all traces. In a production
+    // application, use `Sampler::Parent` or `Sampler::Probability` with a desired probability.
+    let provider = sdk::Provider::builder()
+        .with_simple_exporter(exporter)
+        .with_config(sdk::Config {
+            default_sampler: Box::new(Sampler::Always),
+            ..Default::default()
+        })
+        .build();
+    let tracer = provider.get_tracer("grpc-server");
+
+    let opentelemetry = tracing_opentelemetry::layer().with_tracer(tracer);
+    tracing_subscriber::registry()
+        .with(opentelemetry)
+        .try_init()?;
+
+    Ok(())
+}
+
+struct HttpHeaderMapCarrier<'a>(&'a tonic::metadata::MetadataMap);
+impl<'a> api::Carrier for HttpHeaderMapCarrier<'a> {
+    fn get(&self, key: &'static str) -> Option<&str> {
+        self.0
+            .get(key.to_lowercase().as_str())
+            .and_then(|value| value.to_str().ok())
+    }
+
+    fn set(&mut self, _key: &'static str, _value: String) {
+        unimplemented!()
+    }
+}
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    tracing_init()?;
+    let addr = "[::1]:50051".parse()?;
+    let greeter = MyGreeter::default();
+
+    Server::builder()
+        .add_service(GreeterServer::new(greeter))
+        .serve(addr)
+        .await?;
+
+    Ok(())
+}