+++ opentelemetry-zipkin/src/lib.rs
@@ -22,7 +22,7 @@
 //!     let provider = sdk::Provider::builder()
 //!         .with_simple_exporter(exporter)
 //!         .with_config(sdk::Config {
-//!             default_sampler: Box::new(sdk::Sampler::Always),
+//!             default_sampler: Box::new(sdk::Sampler::AlwaysOn),
 //!             ..Default::default()
 //!         })
 //!         .build();