+++ opentelemetry-jaeger/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.5.0
+
+### Changed
+- Update to opentelemetry v0.6.0
+
+### Fixed
+- Switch internally to `ureq` from `reqwest` to fix #106
+- Fix exported link span id format #118
+
 ## v0.4.0
 
 ### Added